async_api/tests/functional/requirements.txt
pytest==6.1.2
pytest-asyncio==0.12.0
aiohttp==3.7.2
elasticsearch[async]==7.9.1
asyncio==3.4.3
pydantic==1.9.0
python-dotenv==0.20.0